requirements.txt
matplotlib==3.5.0
pandas==1.3.4
Pillow==11.1.0
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
tqdm==4.67.1
typing_extensions==4.12.2
